pyproject.toml
[tool.poetry]
name = "tutorialpress"
version = "0.1.0"
description = ""
authors = ["Natan"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.3"
djangorestframework = "^3.11.0"
drf-yasg = "^1.17.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.1.1"
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"

[tool.black]
line-length = 119

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

